Adding path definitions to the imported files (#1)
Hi Tagnaouti, Sorry but I missed this one. I'm about to have a look at another round of integrations. And there is some feedback on the main bolt compiler that covers how we intend on handling paths.

Watch this space, I haven't had time to fully integrate the compiler into the root project but we will more than likely be dropping the Gulp plugin in favor of the upstream compiler in the future.

Cheers,
Jason